requirements.txt
pandas>=1.5.0
numpy>=1.21.0
xgboost>=1.6.0
scikit-learn>=1.1.0
google-play-scraper>=1.2.0
itunes-app-scraper-dmi>=0.9.6
thefuzz>=0.19.0
matplotlib>=3.5.0
seaborn>=0.11.0
joblib>=1.1.0
requests>=2.28.0
urllib3>=1.26.0
six>=1.16.0
